Log Screen Changes
- Log screen now displays training data by month
- Forward and back buttons now change data based on relation to current
month (i.e. pressing back 4 times will set an integer to -4, this is
then used to query the database for data between start of that month and
the end of it)
- Implementation is slightly buggy because SQLite implementation of
dates (specifically using something like '-1 month')
- From official documentation, "Note that "±NNN months" works by
rendering the original date into the YYYY-MM-DD format, adding the ±NNN
to the MM month value, then normalizing the result. Thus, for example,
the data 2001-03-31 modified by '+1 month' initially yields 2001-04-31,
but April only has 30 days so the date is normalized to 2001-05-01."
- Plan is to use java's calendar.add() to get a more precise solution to
changing months